fix; refactor components in folder
