+++ src/components/client/PackagePayment.js
@@ -1,8 +1,8 @@
 import axios from "axios";
 import React, { useEffect, useState } from "react";
 import { useHistory, useParams } from "react-router-dom";
-import { url } from "../api/api";
-import PayBill from "../utils/PayBill";
+import { url } from "../../api/api";
+import PayBill from "../../utils/PayBill";
 
 const PackagePayment = () => {
   const history = useHistory();
